Clarify lifecycle event selection process
Previously, the name "selection" described both a D3.js selection *and*
a function that returned a D3.js selection. This organization reflects
the distinction more clearly. (It also avoids the overhead of creating a
new function via `Function#bind`.)

Add comments to more explicitly explain the intent of the code.